pyproject.toml
[project]
name = "tomtoolkit"  # must match PyPI package name
description = "TOM Toolkit and base modules"
authors = [
    {name = "TOM Toolkit Project", email = "tomtoolkit-maintainers@lco.global"},
]
maintainers = [
    {name = "Joey Chatelain", email = "jchate6@gmail.com"},
    {name = "William Lindstrom", email = "llindstrom@lco.global"},
    {name = "Rachel Street", email = "rstreet@lco.global"},
]
license = "GPL-3.0-only"
readme = "README.md"
homepage = "https://lco.global/tomtoolkit/"
repository = "https://github.com/TOMToolkit/tom_base"
documentation = "https://tom-toolkit.readthedocs.io/en/latest/"
keywords = [
    'tomtoolkit',
    'TOM',
    'astronomy',
    'astrophysics',
    'cosmology',
    'science',
    'fits',
    'observatory',
]
classifiers = [
    'Development Status :: 3 - Alpha',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: BSD License',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Topic :: Scientific/Engineering :: Astronomy',
    'Topic :: Scientific/Engineering :: Physics'
]

dynamic = ["version"]
requires-python = ">=3.9.0,<3.13"
dependencies = [
    "numpy >1.24,<2.2",
    "astroquery <1",
    "astroplan >=0.8,<0.11",
    "astropy >=5.3.3,<7 ; python_version >= '3.10'",
    "astropy >=5.3.3,<6 ; python_version < '3.10'",
    "cryptography <=46",
    "django <5",
    "djangorestframework >=3.15,<4",
    "django-bootstrap4 >3,<25",
    "beautifulsoup4 <5",
    "django-contrib-comments >=2.0,<3.0",
    "django-crispy-forms >=2.0,<3.0",
    "crispy-bootstrap4 >=2023.1,<2024.11",
    "django-extensions >=3.0.0,<4.0.0",
    "django-filter >=23,<25",
    "django-gravatar2 >=1.0.0,<2.0.0",
    "django-guardian >=2.0.0,<3.0.0",
    "django-htmx >=1.0.0,<2.0.0",
    "fits2image >=0.4,<0.5",
    "markdown <4",
    "pillow >9.2,<12.0",
    "plotly <6",
    "python-dateutil <3",
    "requests <3",
    "specutils <2",
    "importlib-resources >=6.4.5,<6.5.0",
    "django-tasks (>=0.6.1,<0.7.0)",
]

[tool.poetry]
packages = [ # tell poetry build that the repo is named tom_base (which != name, above)
    { include="tom_alerts"},
    { include="tom_base"},
    { include="tom_catalogs"},
    { include="tom_common"},
    { include="tom_dataproducts"},
    { include="tom_dataservices"},
    { include="tom_observations"},
    { include="tom_setup"},
    { include="tom_targets"},
    { include="*", format="sdist"},
]
version = "0.0.0" # version supplied by poetry-dynamic-versioning

[tool.poetry.group.test.dependencies]
responses = ">=0.23,<0.26"
factory_boy = ">3.2.1,<3.4"
psycopg2-binary = "*" # for testing postgres

[tool.poetry.group.docs.dependencies]
docutils = "!=0.21.post1"
recommonmark = "~0.7"
sphinx = ">=5,<8"
sphinx-design = ">=0.6,<0.7"
sphinx-rtd-theme = ">=1.0,<3.1"
sphinx-copybutton = "~0.5"

[tool.poetry.group.coverage.dependencies]
coverage = ">=6,<8" # coveralls needs ~6 even though 7.3.2 is latest
coveralls = ">=3,<5"

[tool.poetry.group.lint.dependencies]
flake8 = ">=6.0,<7.2"


[tool.poetry.requires-plugins]
poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"
# the default pattern regex makes the 'v' manditory
# this pattern modifies the default regex in order to make the 'v' optional
# ('v' becomes '[v]?' meaning a single v, [v], and ? means optional)
pattern = "(?x)^[v]?((?P<epoch>\\d+)!)?(?P<base>\\d+(\\.\\d+)*)([-._]?((?P<stage>[a-zA-Z]+)[-._]?(?P<revision>\\d+)?))?(\\+(?P<tagged_metadata>.+))?$"

# substitute version not only in pyproject.toml (which the config block above does)
# but also the __version__.py file (using the default value of the files property).
[tool.poetry-dynamic-versioning.substitution]


[build-system]
requires = ["poetry-core >=1.0.0", "poetry-dynamic-versioning >=1.0.0, <2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
